modified:   client/package-lock.json 	modified:   client/src/components/CreateDeal.jsx 	modified:   client/src/components/Deals/Deal/Deal.jsx 	modified:   client/src/components/Deals/Deals.jsx 	modified:   client/src/components/Navbar/Favorite/Favorite.jsx 	modified:   client/src/components/Profile.jsx 	modified:   client/src/components/UserDeal.jsx 	modified:   client/src/features/api/apiSlice.js 	modified:   client/src/features/api/dealSlice.js 	modified:   server/controllers/deals.js 	modified:   server/controllers/user.js 	modified:   server/index.js 	modified:   server/routes/users.js 	new file:   server/routes/usersUpdate.js

+++ server/controllers/deals.js
@@ -1,6 +1,7 @@
 const { StatusCodes } = require("http-status-codes");
 const Deal = require("../models/Deals");
 const { NotFoundError, BadRequestError } = require("../errors");
+const { default: mongoose } = require("mongoose");
 
 const getPostsBySearch = async (req, res) => {
   const { searchQuery } = req.query;
@@ -42,16 +43,20 @@ const getUserDeal = async (req, res) => {
 const deleteDeal = async (req, res) => {
   const {
     user: { userId },
-    params: { id: dealId },
+    params: { id: _id },
   } = req;
-  const deal = await Deal.findByIdAndDelete({
-    _id: dealId,
-    createdBy: userId,
-  });
-  if (!deal) {
-    throw new NotFoundError(`Kein Deal mit id: ${dealId}`);
+  console.log(userId, _id);
+  if (!mongoose.Types.ObjectId.isValid(userId)) return res.sendStatus(404);
+  const findDeal = await Deal.findOne({ _id, createdBy: userId });
+  if (!findDeal) {
+    throw new NotFoundError(`Kein Deal mit id: ${_id} gefunden`);
   }
-  res.status(StatusCodes.OK).send();
+  if (findDeal) {
+    await Deal.findByIdAndDelete({
+      _id,
+    });
+  }
+  res.status(StatusCodes.OK).json({ _id });
 };
 
 const updateDeal = async (req, res) => {
